GeminiRAG/requirements.txt
google-genai
google-adk
fastapi
uvicorn
streamlit
python-dotenv
requests
rank-bm25==0.2.2